Rename StopsFinder to StopFinder
