+++ src/doc/examples/ol-stop-finder.html
